requirements.txt
Flask==3.1.0
pandas==2.2.3
gunicorn==21.2.0
PyMySQL==1.1.1
boto3==1.37.14
python-dotenv==0.21.0